+++ lib/jalc/registration/client.rb
@@ -9,22 +9,19 @@ require_relative 'middleware/raise_error'
 
 module JaLC
   module Registration
-    BASE_URL = 'https://japanlinkcenter.org'
-
     class Client
-      def initialize(id:, password:, logger: nil, base_url: BASE_URL)
-        @id = id
-        @password = password
-        @logger = logger
-        @base_url = base_url
+      attr_reader :config
+
+      def initialize(config)
+        @config = config
       end
 
       def post(xml_file)
         response = conn.post(
           '/jalc/infoRegistry/registDataReceive/index',
           {
-            login_id: Faraday::Multipart::ParamPart.new(@id, 'text/plain'),
-            login_password: Faraday::Multipart::ParamPart.new(@password, 'text/plain'),
+            login_id: Faraday::Multipart::ParamPart.new(config.id, 'text/plain'),
+            login_password: Faraday::Multipart::ParamPart.new(config.password, 'text/plain'),
             fname: Faraday::Multipart::FilePart.new(xml_file, 'text/xml'),
           },
         )
@@ -35,8 +32,8 @@ module JaLC
         response = conn.get(
           '/jalc/infoRegistry/registDataResult/index',
           {
-            login_id: @id,
-            login_password: @password,
+            login_id: config.id,
+            login_password: config.password,
             exec_id: exec_id,
           },
         )
@@ -47,15 +44,15 @@ module JaLC
 
       def conn
         @conn ||= Faraday.new(
-          url: @base_url,
+          url: config.base_url,
           headers: { 'User-Agent' => "jalc-ruby v#{VERSION}" },
         ) do |f|
           f.request :multipart
           f.use Middleware::RaiseError
           f.use Middleware::ParseXML
           f.response :raise_error
-          if @logger
-            f.response :logger, @logger, { headers: false } do |logger|
+          if config.logger
+            f.response :logger, config.logger, { headers: false } do |logger|
               logger.filter(/(password=)\w+/, '\1[FILTERED]')
             end
           end