+++ lib/jalc/registration/config.rb
@@ -5,9 +5,12 @@ require 'logger'
 module JaLC
   module Registration
     class Config
-      attr_accessor :id, :password, :logger
+      DEFAULT_BASE_URL = 'https://japanlinkcenter.org'
+
+      attr_accessor :base_url, :id, :password, :logger
 
       def initialize
+        @base_url = DEFAULT_BASE_URL
         @id = nil
         @password = nil
         @logger = ::Logger.new($stdout)