+++ lib/jalc/rest/config.rb
@@ -5,9 +5,12 @@ require 'logger'
 module JaLC
   module REST
     class Config
-      attr_accessor :logger
+      DEFAULT_BASE_URL = 'https://api.japanlinkcenter.org'
+
+      attr_accessor :base_url, :logger
 
       def initialize
+        @base_url = DEFAULT_BASE_URL
         @logger = ::Logger.new($stdout)
       end
     end